+++ tech1-framework-b2b-base-security-jwt/src/main/java/io/tech1/framework/b2b/base/security/jwt/crons/SessionsCron.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
+@SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection")
 @Slf4j
 @Service
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
@@ -29,6 +30,18 @@ public class SessionsCron {
             this.sessionRegistry.cleanByExpiredRefreshTokens(usernames);
         } else {
             LOGGER.debug("Sessions cleanup by expired JWT refresh tokens is disabled");
+        }
+    }
+
+    @Scheduled(
+            cron = "${tech1.securityJwtConfigs.sessionConfigs.enableSessionsMetadataRenewCron.expression}",
+            zone = "${tech1.securityJwtConfigs.sessionConfigs.cleanSessionsByExpiredRefreshTokensCron.zoneId}"
+    )
+    public void enableSessionsMetadataRenew() {
+        if (this.applicationFrameworkProperties.getSecurityJwtConfigs().getSessionConfigs().getEnableSessionsMetadataRenewCron().isEnabled()) {
+            // WARNING: add session registry business method
+        } else {
+            LOGGER.debug("Sessions renew metadata cron is disabled");
         }
     }
 }