+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 - Docker "servers" (tech1-framework-b2b-mongodb-server, tech1-framework-b2b-postgres-server) → GitHub Packages
 - Add "create_at", "updated_at", "metadata_renew_cron", "metadata_renew_manually" → tech1_users_sessions (PostgreSQL) + liquibase
 - Add "create_at", "updated_at", "metadata_renew_cron", "metadata_renew_manually" → tech1_users_sessions (MongoDB) + MIGRATION
+- Add "enableSessionsMetadataRenewCron" configuration
 
 ### v1.15
 - Move <distributionManagement> ossrh → "release" profile